shortbread: Limit max water labels
At low zooms there can be thousands of water labels in one tile.
These correspond to water features that are too tiny to make it
into the tile, and there's no way to fit thousands of labels into
one tile. This limits it to 256 labels for zoom < 14. 256 is too
many to label, but the client can handle label collision.