Port Street Spirit to glug
MapLibre Styles are not designed to be directly written, so some
pre-processor needs to be used. Previously charites was used,
but it had limitations on includes that dictated the project
structure instead of grouping more than one layer in a file.

Instead, the style has been ported to glug. glug is a Ruby DSL
for MapLibre styles. It is less awkward to write and is powerful
enough to do what the style needs.

Because glug is ruby-based, the ability to lint and format code
with rubocop is added. This uses custom rubocop rules that make
sense for stylesheets.

Currently the Gemfile points at a fork of Glug as it relies on
some bugfixes that have not yet been merged upstream.

Sprite generation has advanced and spreet is now used to generate
spritesheets instead of basemaps/sprites.

With the software changes INSTALL.md was split to have separate
directions for shortbread.